🎁 Add `ValkyrieWorkGenerator` logic
This generator is a simple service that can create Valkyrie works
typically used in development when you want to test out work creation.

It can be as simple as:

```rb
Skullrax::ValkyrieWorkGenerator.new.create
```

or a little more customized like:

```rb
generator = Skullrax::ValkyrieWorkGenerator.new(
  model: Monograph,
  user: current_user,
  title: ['Sample Work Title'],
  keyword: ['sample', 'work', 'keywords'],
  visibility: 'open'
)

generator.create
```

It will fill out any missing required fields with placeholder data such
as "Test title".  If a required field is a controlled vocabulary through
the questioning authority gem, then it will do it's best to sniff out an
appropriate value.

If the work fails to save the you can check `generator.errors` to see
if an error was caught.  If the error occurs before the save action then
it should be raised as usual.