🎁 Handle Hyku's question authority oddities
In Hyku there are some authorities that aren't named as expected for
example 'audience.yml' should really have been 'audiences.yml' to match
the pattern of other authorities.  This commit updates the
ControlledVocabularyHandler to first try the pluralized form of the
property name, and if that fails, to try the singular form.  This allows
us to handle these oddities.