🎁 Add a lookup to handle geonames property
This commit will add a handler to look up geonames URIs so the user
doesn't need to have the exact URI on hand when adding a geonames
property to a work.

Usage example:

```rb
Skullrax::ValkyrieWorkGenerator.new(
  geonames: 'san diego'
).generate
```

This will search geonames.org for "san diego" and use the first result's
URI (which happens ot be `https://sws.geonames.org/5391811/`) as the
value for the geonames property.