♻️ Refactor transaction executor
I was trying to consolidate the transaction logic from the collection
generator into the transaction executor but I think they're different
enough where it doesn't make sense to share that code.  I've renamed
the transaction executor to WorkTransactionExecutor to make it clear
that it's only for works and did some slight refactoring.